extracted demo into its own file
