+++ ml-intro/imgui.ini
@@ -13,3 +13,26 @@ Pos=650,20
 Size=550,680
 Collapsed=0
 
+[Window][Example: Property editor]
+Pos=60,60
+Size=430,450
+Collapsed=0
+
+[Window][Example: Property editor/##tree_73339443]
+IsChild=1
+Size=300,415
+
+[Window][Example: Simple layout]
+Pos=60,60
+Size=500,440
+Collapsed=0
+
+[Window][Example: Simple layout/left pane_AED60EF8]
+IsChild=1
+Size=229,386
+
+[Table][0xAC9A836A,2]
+RefScale=13
+Column 0  Width=4
+Column 1  Weight=2.0000
+
